Fix Catalina Unus integration

+++ resources/patch.sh
@@ -375,7 +375,9 @@ Patch_Unus()
 	if [[ $catalina_unus == "1" ]]; then
 		echo -e $(date "+%b %m %H:%M:%S") ${text_progress}"> Patching Catalina Unus."${erase_style}
 
-			Output_Off rm -r "$volume_path"/System/Library/Templates/Data/private/var/db/dslocal
+			if [[ -d "$volume_path"/private/var/db/dslocal ]]; then
+				Output_Off rm -r "$volume_path"/System/Library/Templates/Data/private/var/db/dslocal
+			fi
 			ditto "$volume_path"/System/Library/Templates/Data "$volume_path"/
 
 		echo -e $(date "+%b %m %H:%M:%S") ${move_up}${erase_line}${text_success}"+ Patched Catalina Unus."${erase_style}
